+++ packages/picasso.js/src/core/chart-components/axis/axis.js
@@ -1,50 +1,13 @@
 import extend from 'extend';
 
 import nodeBuilder from './axis-node-builder';
-import { discreteDefaultSettings, continuousDefaultSettings } from './axis-default-settings';
+import {
+  DEFAULT_CONTINUOUS_SETTINGS,
+  DEFAULT_DISCRETE_SETTINGS
+} from './axis-default-settings';
 import calcRequiredSize from './axis-size-calculator';
-import resolveSettingsForPath from '../settings-setup';
-import { resolveForDataValues } from '../../style';
 import crispify from '../../transposer/crispifier';
 import { scaleWithSize } from '../../scales';
-
-/**
- * @typedef {object} component--axis
- *
- */
-
-/**
- * @typedef {object} component--axis.settings
- * @property {object} [labels] - Labels settings
- * @property {boolean} [labels.show=true]
- * @property {string} [labels.mode='auto'] - Control how labels arrange themself. Availabe modes are auto, horizontal, layered and tilted. Only horizontal is supported on a continuous axis
- * @property {number} [labels.tiltAngle=40] - Angle in degrees, capped between -90 and 90
- * @property {number} [labels.maxEdgeBleed=Infinity]
- * @property {string} [labels.fontFamily='Arial']
- * @property {string} [labels.fontSize='12px']
- * @property {string} [labels.fill='#595959']
- * @property {number} [labels.margin] - Space between tick and label. Default to 6 (discrete) or 4 (continuous)
- * @property {number} [labels.maxLengthPx=150] - Max length of labels in pixels
- * @property {number} [labels.minLengthPx=0] - Min length of labels in pixels. Labels will always at least require this much space
- * @property {number} [labels.maxGlyphCount=NaN] - When only a sub-set of data is available, ex. when paging. This property can be used to let the axis estimate how much space the labels will consume, allowing it to give a consistent space estimate over the entire dataset when paging.
- * @property {number} [labels.align=0.5] - Align act as a slider for the text bounding rect over the item bandwidth, given that the item have a bandwidth. Except when labels are tilted, then the align is a pure align that shifts the position of the label anchoring point.
- * @property {object} [line]
- * @property {boolean} [line.show=true]
- * @property {number} [line.strokeWidth=1]
- * @property {string} [line.stroke='#cccccc']
- * @property {object} [ticks]
- * @property {boolean} [ticks.show=true]
- * @property {number} [ticks.margin=0]
- * @property {number} [ticks.tickSize] - Default to 4 (discrete) or 8 (continuous)
- * @property {string} [ticks.stroke='#cccccc']
- * @property {number} [ticks.strokeWidth=1]
- * @property {object} [minorTicks] - Only on a continuous axis
- * @property {boolean} [minorTicks.show=true]
- * @property {number} [minorTicks.margin=0]
- * @property {number} [minorTicks.tickSize=3]
- * @property {string} [minorTicks.stroke='#e6e6e6']
- * @property {number} [minorTicks.strokeWidth=1]
- */
 
 function alignTransform({ align, inner }) {
   if (align === 'left') {
@@ -66,12 +29,24 @@ function resolveAlign(align, dock) {
   return dock; // Invalid align, return current dock as default
 }
 
-function resolveInitialStyle(settings, baseStyles, chart) {
-  const ret = {};
-  Object.keys(baseStyles).forEach((s) => {
-    ret[s] = resolveSettingsForPath(settings, baseStyles, chart, s);
-  });
-  return ret;
+/**
+ * @ignore
+ * @param {object} context - The component context
+ */
+function resolveLocalSettings({
+    state,
+    style,
+    settings
+  }) {
+  const defaultStgns = extend(true, {}, state.isDiscrete ? DEFAULT_DISCRETE_SETTINGS : DEFAULT_CONTINUOUS_SETTINGS, style);
+  const localStgns = extend(true, {}, defaultStgns, settings.settings);
+
+  const dock = settings.dock || state.defaultDock;
+  localStgns.dock = dock;
+  localStgns.align = resolveAlign(settings.settings.align, dock);
+  localStgns.labels.tiltAngle = Math.max(-90, Math.min(localStgns.labels.tiltAngle, 90));
+
+  return localStgns;
 }
 
 function updateActiveMode(state, settings, isDiscrete) {
@@ -94,8 +69,13 @@ const axisComponent = {
   defaultSettings: {
     displayOrder: 0,
     prioOrder: 0,
-    paddingStart: 0,
-    paddingEnd: 10
+    settings: {},
+    style: {
+      labels: '$label',
+      ticks: '$guide-line',
+      minorTicks: '$minor-guide-line',
+      line: '$guide-line'
+    }
   },
   created() {
     // State is a representation of properties that are private to this component defintion and may be modified by only in this context.
@@ -108,39 +88,26 @@ const axisComponent = {
       ticks: [],
       innerRect: { width: 0, height: 0, x: 0, y: 0 },
       outerRect: { width: 0, height: 0, x: 0, y: 0 },
-      defaultStyleSettings: undefined,
       defaultDock: undefined,
       concreteNodeBuilder: undefined,
       settings: undefined
     };
 
     if (this.state.isDiscrete) {
-      this.state.defaultStyleSettings = discreteDefaultSettings();
       this.state.defaultDock = 'bottom';
     } else {
-      this.state.defaultStyleSettings = continuousDefaultSettings();
       this.state.defaultDock = 'left';
     }
 
     this.setState(this.settings);
   },
-  setState(settings) {
+  setState() {
     this.state.isDiscrete = !!this.scale.bandwidth;
-    const styleSettings = resolveInitialStyle(settings.settings, this.state.defaultStyleSettings, this.chart);
-    this.state.settings = extend(true, {}, this.settings, settings.settings, styleSettings);
-
-    const dock = typeof this.state.settings.dock !== 'undefined' ? this.state.settings.dock : this.state.defaultDock;
+    this.state.settings = resolveLocalSettings(this);
 
     this.state.concreteNodeBuilder = nodeBuilder(this.state.isDiscrete);
 
-    this.state.settings.dock = dock;
-    this.state.settings.align = resolveAlign(this.state.settings.align, dock);
-    this.state.settings.labels.tiltAngle = Math.max(-90, Math.min(this.state.settings.labels.tiltAngle, 90));
-    this.dockConfig.dock = dock; // Override the dock setting (TODO should be removed)
-
-    Object.keys(styleSettings).forEach((a) => {
-      this.state.settings[a] = resolveForDataValues(this.state.settings[a]);
-    });
+    this.dockConfig.dock = this.state.settings.dock; // Override the dock setting (TODO should be removed)
 
     this.state.isHorizontal = this.state.settings.align === 'top' || this.state.settings.align === 'bottom';
     this.state.labels.activeMode = updateActiveMode(this.state, this.state.settings, this.state.isDiscrete);