+++ packages/picasso.js/src/core/chart-components/axis/index.js
@@ -1,5 +1,16 @@
 import axisComponent from './axis';
 
+/**
+ * @typedef {object} component--axis
+ */
+
+/**
+ * @type {string}
+ * @memberof component--axis
+ */
+
+const type = 'axis';
+
 export default function axis(picasso) {
-  picasso.component('axis', axisComponent);
+  picasso.component(type, axisComponent);
 }